+++ private/REMARK_STRINGIFY_OPTIONS.js
@@ -1,0 +1,15 @@
+'use strict';
+
+/**
+ * [`remark-stringify`](https://npm.im/remark-stringify) `stringify` options for
+ * stringifying a markdown AST as close to Prettier format as possible.
+ * @kind constant
+ * @name REMARK_STRINGIFY_OPTIONS
+ * @ignore
+ */
+module.exports = {
+  bullet: '-',
+  emphasis: '_',
+  listItemIndent: '1',
+  rule: '-',
+};