+++ private/remarkStringifyOptions.js
@@ -1,7 +1,0 @@
-'use strict';
-
-// Prettier formatting.
-exports.bullet = '-';
-exports.emphasis = '_';
-exports.listItemIndent = '1';
-exports.rule = '-';